Monkeypatch pickle generation to use a temp directory.
When multiple processes call load_grammar simultaneously and the pickle doesn't
already exist (such as on CI), they race and one can see a half-written pickle
since it was previously written in its final location.

Writes into a subdirectory that in most cases will be on the same mount as the
destination so atomic renames work.

On Windows, the rename will fail if the destination exists (due to being
created by another process in the meantime).  I think on *nix the rename should
succeed and readers will only see fully-written pickles.