Allow tests to be run out of the source tree
Debian has "autopkgtests" that can run package test suites against the
installed version of the package to prove that it was installed
successfully. They also help to prove that the dependencies remain
compatible.

Upstream unit tests are usually run by running the tests outside the
source tree. This is trivial when tests are in a separate module.

Fissix's tests are within the fissix module. Many imports would work
correctly outside source tree, but relative imports between the test and
source packages need to be made explicit.